Rename constants folder to types
