+++ js/viz/draw/app.js
@@ -3,7 +3,7 @@ import { Viewport } from "pixi-viewport";
 import { dragEnd } from "./drag.js";
 import { setRenderable } from "./renderable.js";
 import { getPixiState } from "../pixi.js";
-import { objectColor } from "../constants/vizStyles.js";
+import { objectColor } from "../types/vizStyles.js";
 
 const pixi = getPixiState();
