+++ js/viz/draw/font.js
@@ -1,5 +1,5 @@
 import { HTMLText, TextStyle } from "pixi.js";
-import { objectColor, size } from "../constants/vizStyles.js";
+import { objectColor, size } from "../types/vizStyles.js";
 
 const MARGIN = 16;
 const PADDING = 8;
