+++ js/viz/types/particleMappings.js
