+++ js/viz/types/simStatus.js
